Make abstract command manager and refactor tab complete
